Update talker.c
Once you have done that, modify the code so that it prints the the IP and port that is associated with the socket (sockfd). Modify it so that the information is printed before the sendto(), and after sendto(). Is there any difference in the output? Do you have any idea why?

In the next step, we will modify the client to accept multiple messages and send each as a separate UDP message.